Replace maven build with a much better and more flexible ant build
Tried adding mockrunner-0.4.2 to the build, but unfortunatley it does not have
any maven repository. So the obvious thing to do would be to just create a
"lib/test" directory, put mockrunner there and tell maven to include it in
the classpath for test. But with maven absolutley nothing is obvious.
There simply is no sane way to to it, or two hours of googling and reading
the poor maven documentation is not enough. Dunno.
What could be done is to add the libs to the maven repo like this:
	mvn install:install-file
	  -Dfile=<path-to-file>
	  -DgroupId=<group-id>
	  -DartifactId=<artifact-id>
	  -Dversion=<version>
	  -Dpackaging=<packaging>
	  -DgeneratePom=true
But that is just insane. The build is supposed to be "git clone; <build tool>",
but with maven this seems impossible. Everything is non-obvious, the documentation
is poor and this "download everyhing on first use" thing is not helpful either.

An ant build on the other hand is simple, obvious and gets to the point. Also ant
is documentet very well. Duplicating build files is a problem, so is adding each
library to each project. But it has to be done just once. And the added size to
the repo does not hurt either.

A tool is something that should make work easier and help getting a task done.
Ant does that, git does it, vim does it. Maven never has and never will because
it is an overengineered piece of bulls***. The basic idea behind it is nice,
but the rest of the concept and the implementation is just bloody brain dead.

In the end, the build has to be fully automated. That's all that matters. It
really does not matter at all if it's maven or ant, as long as it just works.

-- D.Ritz who will never use maven again.

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>