Add Filter bases security, login
This adds a filter based security system, described in the architecture doc.
Also adds the login page, some useful Struts helpers and tests for the
security stuff. Tests for the Forms/Actions are still missing...need to
read some manuals first.

The whole security filtering and struts stuff is mostly imported from
another project. The description in the architecure doc and the mockrunner
based test cases are new.

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>